fix : random tour serializer field 수정

+++ random_tour/serializers.py
@@ -13,8 +13,8 @@ class SaveRandomTourOrderSerializer(serializers.ModelSerializer):
     tour_id = serializers.IntegerField(source='tour.id')
     tour_name = serializers.CharField(source='tour.title')
     tour_address = serializers.CharField(source='tour.addr1')
-    tour_mapx = serializers.IntegerField(source='tour.mapx')
-    tour_mapy = serializers.IntegerField(source='tour.mapy')
+    tour_mapx = serializers.FloatField(source='tour.mapx')
+    tour_mapy = serializers.FloatField(source='tour.mapy')
 
     class Meta:
         model = RandomTourOrder
@@ -33,8 +33,8 @@ class RandomOrderSerializer(serializers.Serializer):
     tour_id = serializers.IntegerField()
     tour_name = serializers.CharField()
     tour_address = serializers.CharField()
-    tour_mapx = serializers.IntegerField()
-    tour_mapy = serializers.IntegerField()
+    tour_mapx = serializers.FloatField()
+    tour_mapy = serializers.FloatField()
 
 class RandomListSerializer(serializers.Serializer):
     tour_list = RandomOrderSerializer(many=True)
